# docs/api/openapi_spec.yaml
openapi: 3.0.1
info:
  title: Fantasy Football API
  version: v1
  description: >
    API for onboarding and querying fantasy football league data.
    Supports ESPN and Sleeper as data sources. League data is fetched
    from external APIs, transformed, and stored in DynamoDB for fast retrieval.
  license:
    name: MIT
paths:
...
  /leagues/{leagueId}:
    get:
      summary: Get league metadata
      description: >
        Checks whether a league has been onboarded. If the league exists,
        returns the league name, the list of seasons the league has been
        active for, and its subscription_end_time. Returns 404 if the league
        has not been onboarded yet. This endpoint is never subscription-gated.
      operationId: getLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
      responses:
        "200":
          $ref: "#/components/responses/LeagueFoundResponse"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)

    delete:
      summary: Delete league
      description: >
        Deletes an onboarded league and its associated data from DynamoDB and S3.
        Also decrements the total onboarded league count.
      operationId: deleteLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
      responses:
        "200":
          $ref: "#/components/responses/DeleteLeagueResponse"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues:
    post:
      summary: Onboard a new league or refresh an existing league's data
      description: >
        Triggers the onboarding Lambda which fetches all historical season data
        from the specified platform (ESPN or Sleeper), transforms it via DuckDB,
        and writes precomputed views to DynamoDB. The requestType query parameter
        determines whether this is a new onboarding ("ONBOARD") or a refresh of an
        existing league's data ("REFRESH").
      operationId: onboardLeague
      parameters:
        - name: requestType
          in: query
          required: true
          description: > 
            The type of onboarding request being made. Can be "ONBOARD" for new leagues
            or "REFRESH" to refresh data for an existing league.
          schema:
            type: string
            enum: [ONBOARD, REFRESH]
            example: "ONBOARD"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardRequest"
      responses:
        "201":
          $ref: "#/components/responses/OnboardSuccessResponse"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues/{leagueId}/query:
    get:
      summary: Query a precomputed league view
      description: >
        Returns a precomputed data view for the specified league. The queryType
        parameter maps directly to the DynamoDB sort key (e.g. WEEKLY_SCORES#2024).
      operationId: queryLeague
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ESPN or Sleeper league ID
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The name of the platform the fantasy football league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "ESPN"
        - name: queryType
          in: query
          required: true
          description: >
            The precomputed view to retrieve. Season-scoped views require
            a season suffix (e.g. SEASON_STANDINGS#2024). Cross-season views
            such as ALL_TIME_STANDINGS do not require a suffix.
          schema:
            type: string
            # TODO: Fill this out as the list of views evolves
            enum:
              - TEAMS
              - MATCHUPS
              - SEASON_STANDINGS
            example: "SEASON_STANDINGS"
      responses:
        "200":
          $ref: "#/components/responses/QuerySuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
  /leagues/{leagueId}/espn_members:
    post:
      summary: Fetch ESPN league members
      description: >
        Server-side proxy to the ESPN Fantasy API that returns the list of members
        for a given ESPN league. Avoids browser CORS restrictions by making the
        request from the Lambda function.
      operationId: getEspnMembers
      parameters:
        - name: leagueId
          in: path
          required: true
          description: The ID of the current (source) league
          schema:
            type: string
            pattern: '^\d+$'
            example: "12345678"
        - name: platform
          in: query
          required: true
          description: The current platform the league is on
          schema:
            type: string
            enum: [ESPN, SLEEPER]
            example: "SLEEPER"
        - name: espnLeagueId
          in: query
          required: true
          description: The ESPN league ID to fetch members from
          schema:
            type: string
            example: "87654321"
        - name: season
          in: query
          required: true
          description: The ESPN season year
          schema:
            type: string
            example: "2025"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EspnMembersRequest"
      responses:
        "200":
          $ref: "#/components/responses/EspnMembersResponse"
        "402":
          $ref: "#/components/responses/SubscriptionRequiredError"
        "404":
          $ref: "#/components/responses/LeagueNotFoundError"
        "502":
          $ref: "#/components/responses/BadGatewayError"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"
...
  /jobs/{jobId}:
    get:
      summary: Get the status of an onboard/refresh/migrate job
      description: >
        Returns the status of a job triggered by onboarding, refresh, or migration,
        identified by the correlation (job) ID returned when the job was triggered.
        Used for polling until the job completes. When the job fails, a user-friendly
        failure_reason is included. A missing item (never created, or expired after
        its 24h TTL) is reported as FAILED.
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          description: The correlation (job) ID returned when the job was triggered
          schema:
            type: string
            pattern: '^[0-9a-fA-F-]{36}$'
            example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        "200":
          $ref: "#/components/responses/JobStatusSuccessResponse"
        "500":
          $ref: "#/components/responses/InternalError"
      security:
        - ClerkJWT: []
      x-amazon-apigateway-integration:
        type: aws_proxy
        httpMethod: POST        # always POST for Lambda proxy, regardless of your route method
        uri: "arn:aws:apigateway:${aws_region}:lambda:path/2015-03-31/functions/${lambda_arn}/invocations"
        payloadFormatVersion: "2.0"   # use 2.0 for HTTP API (v2)
...
components:
  securitySchemes:
    ClerkJWT:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: jwt
      x-amazon-apigateway-authorizer:
        identitySource: $request.header.Authorization
        jwtConfiguration:
          issuer: "${clerk_issuer_url}"
          audience:
            - "${clerk_jwt_audience}"
        type: jwt
...
  responses:
...
    LeagueFoundResponse:
      description: League found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/APIResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/LeagueFoundData"
          example:
            detail: "Found league"
            data:
              seasons: ["2023", "2024", "2025"]
              league_name: "The Gridiron Gang"
...
    LeagueNotFoundError:
      description: League not found — onboarding required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "League not found"
            message: "No ESPN league with ID 12345678 has been onboarded. Use POST /leagues to onboard."
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Internal server error"
            message: "An unexpected error occurred. Please try again later."
...